requirements.txt
streamlit
youtube-transcript-api
langchain
langchain-community
langchain-groq
langchain-huggingface
faiss-cpu
tiktoken
python-dotenv
sentence-transformers
langchain-google-genai